Create MVC architecture
